requirements.txt
scikit-learn
pandas
pyyaml
joblib
wandb
fastapi
uvicorn